fix ram

+++ jolt-core/src/zkvm/ram/output_check.rs
@@ -53,8 +53,8 @@ struct OutputSumcheckProverState<F: JoltField> {
 impl<F: JoltField> OutputSumcheckProverState<F> {
     #[tracing::instrument(skip_all, name = "OutputSumcheckProverState::initialize")]
     fn new(
-        initial_ram_state: Vec<u32>,
-        final_ram_state: Vec<u32>,
+        initial_ram_state: Vec<u64>,
+        final_ram_state: Vec<u64>,
         program_io: &JoltDevice,
         r_address: &[F],
     ) -> Self {
@@ -129,8 +129,8 @@ pub struct OutputSumcheck<F: JoltField> {
 impl<F: JoltField> OutputSumcheck<F> {
     #[tracing::instrument(skip_all, name = "OutputSumcheck")]
     pub fn new_prover<ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>>(
-        initial_ram_state: Vec<u32>,
-        final_ram_state: Vec<u32>,
+        initial_ram_state: Vec<u64>,
+        final_ram_state: Vec<u64>,
         state_manager: &mut StateManager<'_, F, ProofTranscript, PCS>,
     ) -> Self {
         let (_, _, program_io, _) = state_manager.get_prover_data();
@@ -448,7 +448,7 @@ impl<F: JoltField> ValFinalSumcheck<F> {
     }
 
     pub fn new_verifier<ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>>(
-        initial_ram_state: &[u32],
+        initial_ram_state: &[u64],
         state_manager: &mut StateManager<'_, F, ProofTranscript, PCS>,
     ) -> Self {
         let (_, _, T) = state_manager.get_verifier_data();
